Move code into main module
